requirements.txt
alchemize==0.14.0
asyncpg==0.27.0
aumbry==0.10.0
click==8.1.3
deepmerge==1.1.0
falcon==3.1.1
greenlet==2.0.2
h11==0.14.0
passlib==1.7.4
pike==0.1.1
PyJWT==2.6.0
python-json-logger==2.0.7
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
uvicorn==0.21.1
